refactor(backend): extract experiment crud module
Continues the incremental split of the crud monolith. Experiment moves to
its own module; the router switches to a direct submodule import. Pure
move -- no logic changes.
