+++ apps/backend/src/rhesis/backend/app/crud/__init__.py
@@ -103,29 +103,6 @@ def delete_endpoint(
     )
 
 
-# Experiment CRUD
-def get_experiments(
-    db: Session,
-    skip: int = 0,
-    limit: int = 10,
-    sort_by: str = "created_at",
-    sort_order: str = "desc",
-    filter: str | None = None,
-    organization_id: str = None,
-    user_id: str = None,
-) -> List[models.Experiment]:
-    return (
-        QueryBuilder(db, models.Experiment)
-        .with_related(include(models.Experiment.project))
-        .with_organization_filter(organization_id)
-        .with_visibility_filter(user_id)
-        .with_odata_filter(filter)
-        .with_sorting(sort_by, sort_order)
-        .with_pagination(skip, limit)
-        .all()
-    )
-
-
 # TestSet CRUD
 def get_test_set(
     db: Session, test_set_id: uuid.UUID, organization_id: str = None, user_id: str = None
